Introduce thread-first and thread-last macros
This patch introduces the thread-first and thread-last macros in control-flow.
The macros have been popularized by Clojure and afterwards adopted in Emacs
Lisp (see subr-x.el). This implementation is a port of the latter, where we use
a common thread-iter(ative) tail recursive function for computing the results.
The choice of the full name was made in the spirit of the library,
conservatively avoiding new syntax like -> or ->>.